+++ packages/gasket-plugin-intl/lib/create.js
@@ -14,18 +14,20 @@ export default async function create(gasket, createContext) {
   // adjust config
   gasket.config.intl.locales = ['en-US', 'fr-FR'];
   gasket.config.intl.managerFilename = 'intl.js';
+  gasket.config.intl.nextRouting = false;
   if (createContext.typescript) gasket.config.intl.managerFilename = 'intl.ts';
 
   const { files, pkg, gasketConfig } = createContext;
-  const __filename = fileURLToPath(import.meta.url);
-  const rootDir = new URL('..', pathToFileURL(__filename).toString()).pathname;
+  const fileName = fileURLToPath(import.meta.url);
+  const rootDir = new URL('..', pathToFileURL(fileName).toString()).pathname;
   const isReactProject = pkg.has('dependencies', 'react');
   files.add(`${rootDir}/generator/*`, `${rootDir}/generator/**/*`);
 
   gasketConfig.addPlugin('pluginIntl', name);
 
   const initialConfig = {
-    locales: gasket.config.intl.locales
+    locales: gasket.config.intl.locales,
+    nextRouting: gasket.config.intl.nextRouting
   };
 
   if (createContext.typescript) {