+++ packages/gasket-plugin-intl/CHANGELOG.md
@@ -1,5 +1,18 @@
 # `@gasket/plugin-intl`
 
+## 7.6.4
+
+### Patch Changes
+
+- fdd8860: Avoid **dirname and **filename const names
+  - @gasket/request@7.5.5
+
+## 7.6.3
+
+### Patch Changes
+
+- 217acaa: React 19 & NextJS 16
+
 ## 7.6.2
 
 ### Patch Changes