+++ packages/gasket-template-api-express/CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## 7.0.6
+
+### Patch Changes
+
+- fdd8860: Avoid **dirname and **filename const names
+
+## 7.0.5
+
+### Patch Changes
+
+- 217acaa: React 19 & NextJS 16
+- 4d6c78e: Update lockfile dependencies
+
 ## 7.0.4
 
 ### Patch Changes